requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
mysql-connector-python==8.3.0
pydantic==2.10.5
python-dotenv==1.0.1
pydantic-settings==2.6.1
